Add FreeBSD support
Consolidate Linux and FreeBSD support into the same files using build
tags.

The semantics for mincore are different on FreeBSD than for
Linux (based on the value of vm.mincore_mapped, mincore reports
mappings, not residency). So cut out the TestIsResidentTwoPages for
FreeBSD but continue using it on Linux.

gofmt moved the import line in the test file.